requirements.txt
google-api-python-client>=1.12.8
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
yt-dlp>=2025.1.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.5.0
plotly>=5.0.0
seaborn>=0.13.0
nltk>=3.8
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0
textblob>=0.17.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
sqlparse>=0.4.0
tqdm>=4.66.0
pyyaml>=6.0
Click>=8.1.0
colorama>=0.4.6
customtkinter>=5.2.0
